lots of little work, getting close to being done

+++ salt/modules/solarispkg.py
@@ -1,9 +1,14 @@
 '''
 Package support for Solaris
 '''
+# Todo:
+# Test with states to make sure it works.
+# How to we deal with upgrade?
+# Have it looked over.
 
 import tempfile
 import os
+import shutil
 
 
 def __virtual__():
@@ -96,32 +101,90 @@ def version(name):
 
 def install(name, **kwargs):
     '''
-    Install the passed package
+    Install the passed package. Can install packages from the following sources::
+        
+        * Locally (package already exists on the minion 
+        * HTTP/HTTPS server
+        * FTP server
+        * Salt master
 
-    # pkgadd -n -a /root/admin.file -d /root/testing/ZABagent2.pkg 'all'
-    # Does this work with non-datastream packages?
-
-    Return a dict containing the new package names and versions::
+    Returns a dict containing the new package names and versions::
 
         {'<package>': {'old': '<old-version>',
                    'new': '<new-version>']}
 
-    CLI Example::
+    CLI Example, installing a datastream pkg that already exists on the minion::
 
-        salt '*' pkg.install <package name>
+        salt '*' pkg.install <package name> source=/dir/on/minion/<package name>
+        salt '*' pkg.install gcc-3.3.2-sol10-sparc-local.pkg source=/var/spool/pkg/gcc-3.3.2-sol10-sparc-local.pkg
+
+    CLI Example, installing a datastream pkg that exists on the salt master::
+
+        salt '*' pkg.install <package name> source='salt://srv/prod/<package name>'
+
+    CLI Example, installing a datastream pkg that exists on a HTTP server::
+
+        salt '*' pkg.install <package name> source='http://packages.server.com/<package name>'
+    
+    By default salt automatically provides an adminfile, to automate package installation, with these options set:
+
+        email=
+        instance=quit
+        partial=nocheck
+        runlevel=nocheck
+        idepend=nocheck
+        rdepend=nocheck
+        space=nocheck
+        setuid=nocheck
+        conflict=nocheck
+        action=nocheck
+        basedir=default
+
+    You can override any of these options in two ways. First you can optionally pass any of
+    the options as a kwarg to the module/state to override the default value or you can
+    optionally pass the 'admin_source' option providing your own adminfile to the minions.
+
+    Note: You can find all of the possible options to provide to the adminfile by reading the admin man page:
+
+        man -s 4 admin
+
+    CLI Example - Overriding the 'instance' adminfile option when calling the module directly:
+
+        salt '*' pkg.install <package name> source='salt://srv/salt/<pkgname>' instance="overwrite"  
+
+    CLI Example - Overriding the 'instance' adminfile option when used in a state:
+
+        <pkgname>:
+          pkg.installed:
+            - source: salt://pkgs/<pkgname>
+            - instance: overwrite
+
+    CLI Example - Providing your own adminfile when calling the module directly:
+
+        salt '*' pkg.install <package name> source='salt://srv/salt/<pkgname>' admin_source='salt://srv/salt/adminfile'
+    
+    CLI Example - Providing your own adminfile when using states:
+
+        <pkgname>:
+          pkg.installed:
+            - source: salt://srv/salt/<pkgname>
+            - admin_source: salt://srv/salt/adminfile
     '''
 
-    if 'source' in kwargs:
-         pkgfile = __salt__['file.cache'](kwargs['source'])
+    if not 'source' in kwargs:
+        return 'source option required with solaris pkg installs'
     else:
-         pkgfile = name
+        if (kwargs['source']).startswith('salt://') or (kwargs['source']).startswith('http://') or (kwargs['source']).startswith('https://') or (kwargs['source']).startswith('ftp://'):
+            pkgname = __salt__['cp.cache_file'](kwargs['source'])
+        else:
+            pkgname = (kwargs['source'])
 
     if 'admin_source' in kwargs:
-        adminfile = __salt__['file.cache'](kwargs['admin_source'])
+        adminfile = __salt__['cp.cache_file'](kwargs['admin_source'])
     else:
         # Set the adminfile default variables
         email=kwargs.get('email', '')
-        instance=kwargs.get('instance', 'overwrite')
+        instance=kwargs.get('instance', 'quit')
         partial=kwargs.get('partial', 'nocheck')
         runlevel=kwargs.get('runlevel', 'nocheck')
         idepend=kwargs.get('idepend', 'nocheck')
@@ -153,14 +216,16 @@ def install(name, **kwargs):
     # what got installed.
     old = _get_pkgs()
 
-    cmd = '/usr/sbin/pkgadd -n -a {0} -d {1} \'all\''.format(adminfile, pkgfile)
+    # Install the package
+    cmd = '/usr/sbin/pkgadd -n -a {0} -d {1} \'all\''.format(adminfile, pkgname)
     __salt__['cmd.retcode'](cmd)
 
     # Get a list of the packages again, including newly installed ones.
     new = _get_pkgs()
 
     # Remove the temp adminfile 
-    os.unlink(adminfile)
+    if not 'admin_source' in kwargs:
+        os.unlink(adminfile)
 
     # Return a list of the new package installed.
     return _compare_versions(old, new)
@@ -170,58 +235,88 @@ def remove(name, **kwargs):
     '''
     Remove a single package with pkgrm
 
-    Returns a list containing the removed packages. Since pkgrm on solaris 
-    does not support dependency management. This will always be just one
-    package.
+    By default salt automatically provides an adminfile, to automate package removal, with these options set:
+
+        email=
+        instance=quit
+        partial=nocheck
+        runlevel=nocheck
+        idepend=nocheck
+        rdepend=nocheck
+        space=nocheck
+        setuid=nocheck
+        conflict=nocheck
+        action=nocheck
+        basedir=default
+
+    You can override any of these options in two ways. First you can optionally pass any of
+    the options as a kwarg to the module/state to override the default value or you can
+    optionally pass the 'admin_source' option providing your own adminfile to the minions.
+
+    Note: You can find all of the possible options to provide to the adminfile by reading the admin man page:
+
+        man -s 4 admin
 
     CLI Example::
 
         salt '*' pkg.remove <package name>
+        salt '*' pkg.remove SUNWgit
     '''
 
     # Check to see if the package is installed before we proceed
     if version(name) == '':
         return '' 
 
-    # Set the adminfile default variables
-    email=kwargs.get('email', '')
-    instance=kwargs.get('instance', 'overwrite')
-    partial=kwargs.get('partial', 'nocheck')
-    runlevel=kwargs.get('runlevel', 'nocheck')
-    idepend=kwargs.get('idepend', 'nocheck')
-    rdepend=kwargs.get('rdepend', 'nocheck')
-    space=kwargs.get('space', 'nocheck')
-    setuid=kwargs.get('setuid', 'nocheck')
-    conflict=kwargs.get('conflict', 'nocheck')
-    action=kwargs.get('action', 'nocheck')
-    basedir=kwargs.get('basedir', 'default')
+    if 'admin_source' in kwargs:
+        adminfile = __salt__['cp.cache_file'](kwargs['admin_source'])
+    else:
+        # Set the adminfile default variables
+        email=kwargs.get('email', '')
+        instance=kwargs.get('instance', 'quit')
+        partial=kwargs.get('partial', 'nocheck')
+        runlevel=kwargs.get('runlevel', 'nocheck')
+        idepend=kwargs.get('idepend', 'nocheck')
+        rdepend=kwargs.get('rdepend', 'nocheck')
+        space=kwargs.get('space', 'nocheck')
+        setuid=kwargs.get('setuid', 'nocheck')
+        conflict=kwargs.get('conflict', 'nocheck')
+        action=kwargs.get('action', 'nocheck')
+        basedir=kwargs.get('basedir', 'default')
 
-    # Make tempfile to hold the adminfile contents.
-    fd, adminfile = tempfile.mkstemp(prefix="salt-")
+        # Make tempfile to hold the adminfile contents.
+        fd, adminfile = tempfile.mkstemp(prefix="salt-")
    
-    # Write to file then close it.
-    os.write(fd, "email=%s\n" % email)
-    os.write(fd, "instance=%s\n" % instance)
-    os.write(fd, "partial=%s\n" % partial)
-    os.write(fd, "runlevel=%s\n" % runlevel)
-    os.write(fd, "idepend=%s\n" % idepend)
-    os.write(fd, "rdepend=%s\n" % rdepend)
-    os.write(fd, "space=%s\n" % space)
-    os.write(fd, "setuid=%s\n" % setuid)
-    os.write(fd, "conflict=%s\n" % conflict)
-    os.write(fd, "action=%s\n" % action)
-    os.write(fd, "basedir=%s\n" % basedir)
-    os.close(fd)
+        # Write to file then close it.
+        os.write(fd, "email=%s\n" % email)
+        os.write(fd, "instance=%s\n" % instance)
+        os.write(fd, "partial=%s\n" % partial)
+        os.write(fd, "runlevel=%s\n" % runlevel)
+        os.write(fd, "idepend=%s\n" % idepend)
+        os.write(fd, "rdepend=%s\n" % rdepend)
+        os.write(fd, "space=%s\n" % space)
+        os.write(fd, "setuid=%s\n" % setuid)
+        os.write(fd, "conflict=%s\n" % conflict)
+        os.write(fd, "action=%s\n" % action)
+        os.write(fd, "basedir=%s\n" % basedir)
+        os.close(fd)
+
+    # Get a list of the currently installed pkgs.
+    old = _get_pkgs()
 
     # Remove the package
     cmd = '/usr/sbin/pkgrm -n -a {0} {1}'.format(adminfile, name)
     __salt__['cmd.retcode'](cmd)
 
     # Remove the temp adminfile 
-    os.unlink(adminfile)
+    if not 'admin_source' in kwargs:
+        os.unlink(adminfile)
 
-    # Since pkgrm only removes one pkg at a time just return the package name
-    return [name] 
+    # Get a list of the packages after the uninstall
+    new = _get_pkgs()
+     
+    # Compare the pre and post remove package objects and report the uninstalled pkgs.
+    return _list_removed(old, new)
+
 
 
 def purge(name, **kwargs):
